export `utils` as `plover_build_utils` so it's usable by plugins
