+++ setup.cfg
@@ -44,6 +44,7 @@ packages =
 	plover.machine
 	plover.oslayer
 	plover.system
+	plover_build_utils
 
 [options.entry_points]
 console_scripts =
